Split the mixed tree-panel and ui-state tests into tests/gui
The seven files whose classes all drive real windows already moved under
`tests/gui`. These two were left behind because they were read as mixed,
and a file cannot move whole while a headless class still lives in it.

`tests/test_pilot_tree_panel.py` is genuinely mixed. `MakeMeshInfoTC`
calls the two `MeshInfoTree` class methods on a bare mesh and needs no
window, while `MeshInfoTreeTC`, `EntityTreeWidgetTC`, and `TreePanelTC`
build widgets against the `RManager` singleton and carry the
`NO_LIVE_WINDOW` guard. The three window-driving classes move to
`tests/gui/test_tree_panel.py` and the rest stays. The new file has to
be self-contained, so `_make_sample_mesh` is duplicated rather than
imported across the two directories: a shared import would couple the
fast lane to the GUI lane, which is what this layout exists to avoid.
The helpers only the moved classes used go with them, and the source
file drops its now unused `os` import and `NO_LIVE_WINDOW`.

`tests/test_pilot_ui_state.py` turned out not to be mixed at all. Both
`GeometrySeamTC` and `UiStateTC` carry the `NO_LIVE_WINDOW` guard, and
`FakePart` exists only to feed `UiStateTC`, so the file moves whole to
`tests/gui/test_ui_state.py` with no edit.

Related to #1222
